Add C_init_vars() call from Fortran.
This routine gets called once and will set up the C side of things, so
that C knows about fields, particles, etc, so that we don't have to pass
all that stuff everytime we call from Fortran to C.

Eventually. For now, it just initializes the global C psc struct.