sse2: A major restructuring of the code.
In order to change from single to double precision, we
need to wrap all the intrinsics in macros. The precision is
now controlled with the flag SSE2_DOUBLE in file psc_sse2.h.

This is also a good time to compress the different yz functions
into a single file.

I am now reasonably certain that push_part_yz was failing the current
checks because of precision issues. It would still be bug related, but
the evidence points to floating point error...