sse2: Vectorized the current updating loop.
In order to use vectors here, we need to modify the l{2,3}{min,max} branching
to zero out the coefficients which will not be used by a specific particle.
The extra effort is worth it to do all the flops in parallel here. At best,
we gain a substantial performance boost. At worst, it should run in about the same time
as a serial translation.

Also disabling checking the ouputted currents, as this will fail for single precision.