sse2: Enhancements to field interpolation and naming conventions.
Four important things happen in this commit.

1) Index calculations are now done using integer vector intrinsics.

2) We assume true 2D for the fields to allow for faster access.

3) The file sse2_cgen.h is renamed to the more accurate simd_wrap.h.
It should be straight-forward to add altivec intrinsics to this file
in the future to make it true SIMD. (This commit accidentally sneaked
in during the rebase)

4)Both sse2 and altivec intrinsics seem to use the format:
(register_type)_(action)_(datatype), so I've changed my macros to reflect this.
I find them easier to write this way.