add a counting sort
it appears to be working, but not really optimally yet.

for random particles:
qsort: 751
heapsort: 7375
mergesort: 3890
countsort: 516

for already fully ordered particles:
qsort: 457
heapsort: 6327
mergesort: 93
countsort: 433

for fully ordered particles, counting and copying everything back and
forth is clearly not optimal -- >~ 100 ms are spent on just copying the
out-of-place result back for the counting sort (plus it wastes memory)

Compared to the orginal fortran at ~2600, qsort and countsort are
a substantial improvement, anyway, and pass 1 of the counting sort could
potentially be absorbed into the particle pusher.

Lots of room for further improvement, I think, but doesn't necessarily
come as easy.

(mergesort and heapsort tests were mac os x only)