add countsort2 variant
the only difference is that we're counting on a separate array of cell
indices instead of the cni member of the particles. It's a noticable
performance gain (not counting the actual setup/fill of that separate
array), but nothing great -- the majority of time is spent copying
particles into their new place, and then copying everything back
in-place.