sse2: Implemented an improved particle loading and storing scheme.
Using vector loads/stores along with unpack instructions helps
to prevent unnecessary reads and writes to the stack. On less
sophisticated processors this does not make much of a difference,
but on the fishercat head node (Nehalem/Gainestown) it speeds up
the particle loading by 30-50%. It may be possible to further optimize
to realize this performance increase on older chips.