sse2: Switched to a padding scheme instead of the 'elements' check.
This change was necessary for the new particle loading introduced in the
previous commit. We are now allocating extra memory in the particle
array to pad it out and make {array_size} %VEC_SIZE == 0. Currently the
padding is done in sse2_particles_from_fortran, but when we no longer need
to read from fortran this will have to be done when the local domain is allocated.

It is importatnt to note that we **must** set the mass of the padding particles
to a non-zero value, otherwise we will be NaNs in the momentum pusher.
As long as qni and wni are zero for the padding it will not affect the results.

I've also cleaned up some unused variables. That doesn't really deserve it's own commit.