make some PML code optional
I'm not planning to handle this in C anytime soon, so let's make it
selectable, at least, and if off, let's just stick to E,B and the old
way of doing things...

This disables somewhat more than just PML specific code, in particular
some of the additional time-averaged fields. Should be easy to put them
back if desired, but in general, we need better, configurable I/O for
time-averaged fields, too.

Not writing all of those fields will likely break plotting tools that
have the field layout hardcoded...

Note: Putting ifdefs all over the place really is not the right way, but
being lazy, and planning to redo much of this in C, anyway, it's what it
is.