param: C_init_param_domain(): Use existing values for defaults
Change in plans: Let's not have our own default values for everything,
but keep values that have been previously set (in Fortran or elsewhere)
as defaults, so they can still be changed on the command line. This will
make it easier to set up different test cases without having huge amounts
of parameters on the command line.