bnd: introduce psc_set_n_particles()
Makes it simpler to keep C and Fortran synchronized...
Using this and some more effort, we could get rid of most of the
niloc passing currently happening, but I'd actually rather get rid
of the has-to-be-allocated-from-fortran constraints altogether lately,
which will require passing stuff. So this here is kinda just a bandaid.