init_particles: start generalizing particles / cell
We had just assumed we'll put the same # of particles in every cell,
and then use the weights to match the density. This is how Will did it
for the Harris sheet, but it's not feasible for laser-plasma problems.
(And not clear to me as to whether it's the right thing for the Harris
sheet or not.)

Now this version still isn't doing the same as what the Fortran code did,
but I don't think the Fortran code did do the right thing in the first
place..