testcase: add "langmuir" testcase
This is really more of an example on how to add a testcase, ie.
take an initial condition which exists in Fortran and turn it into
a run-time selectable initial condition.

In this case, we only need to initialize particles, so the new file
"case_langmuir.c" contains really only langmuir_init_nvt(), which sets
density, velocity and temperature at a certain point in space.

Other than that, "case_langmuir.c" needs to be added to src/Makefile.am,
and the testcase needs to be registered in src/psc.h and
src/init_parameters.c