case langmuir: make some parameters command line selectable
This is a demonstration on how to change hard-coded parameters
to parameters which can be selected on the command line --
copy&paste as needed to other initial conditions.

langmuir_create() will create a struct of parameters needed for this
particular case. Defaults will be set as specified, which can be
overridden on the command line.

Other functions, like langmuir_init_nvt() then access those parameters
instead of the hard-coded ones.