cbe: Large port of deprecated cbe work.
This commit brings in all the work done in my cell-alf branch.
The individual commits have been squashed because the bulk of the
work in that branch (ie, struggling with alf) is deprecated and will
not be used. Using ALF was really just an experiment. It turns out to
be good for testing compute kernels, but is largely unsuitable for
PIC. Because it manages so much of the low level SPU functions for
you, it is difficult to fine tune those systems and optimize them for
the problem at hand. The cell-alf branch may evolve into a testing ground
for compute kernels, as it gives some very nice debugging information.

Most of the code in this commit will be trashed, but there are a couple
scattered lines here and there which will carry over (as well as the bulk
of the compute kernel) and thus is it easier to bring this all over and go
from there.

Please avoid branching off this commit, and I will not support anything
but the most recent commit of the cell-dev branch until we decide to
branch for release.