sse2: Several bug fixes and precision enhancements.
In calculating fnqs{}, I was multiplying instead of dividing.

Default precision has been switched to double at the moment.

In a few places, using /root instead of *(1/root) seems to improve agreement with the fortran code.

At this point, the sse2 pusher should be able to pass particle checks against the fortran
down to about double precision machine epsilon. The current checks will only pass down to two
orders of magnitude or so above machine epsilon. This might be indicative of another bug, but
it could also just be compounded round-off error.

I have also made some changes to the sse2 backend functions anticipating some future advancements
in the code. (Notably, the number of particles in the local domain changing.)