doc: Improvements to the script psc_doxyfy.sh
*Won't spawn new xterm window if emacs or xemacs are found.
Spawning a new xterm is slow over ssh. If emacs or xemacs is found
warnings are outputed to a temp file which is opened along with the
file to edit in a console {x}emacs.

It is now possible to skip individual files in editing.

The conf file psc_warn.conf was an unecessary burden, as doxygen can
read from stdin. The build conf file is now run through awk to change
the important variables to gather warning information.