clean: get rid of PIC_push_part_{z,yz}.F
At some point it was useful to separate PIC_push_part_* and
PIC_move_part_* because the former wouldn't pull in modules, but
that's not the case anymore, so we can just put both functions back
PIC_move_part_*, which is how the xz case was done, too.

No actual changes other than moving files and some updated comments.