base: rename FF3 -> F3_BASE
The FF3 macro is used in the C code to access Fortran-layout fields.
In the future, the base data structures will change to not always
be based on Fortran, so we use a generic macro name F3_BASE that will
access whatever data structure the underlying fields are stored in.