sse2: *BUG FIX* Vector integer multiplication is now working correctly.
There is no intrinsic that maps directly from int * int = int. Previously we were
using _mm_mul_epi16, which was not correct at all. We're now using a function to perform
the needed steps for 32bit -> 32bit integer multiplication.

There is some question about the cost of this function. Unit tests,
and a straight adding of latencies, seem to indicate that it is
painfully slow. Searching for a better way...