output: hdf5: write xdmf metadata
This will write metadata so that Paraview can read the HDF5 files.
It only works properly if "--output_combine" is not used, which is
a band-aid for certain cases only, anyway.

Currently, we're not using hyperslabs, but separate blocks, since
hyperslabs make Paraview 3.8.0 crash.

The other surprising thing is that I didn't have to transpose the
HDF5 data to be in C order -- anyway, if it works, why question it...