output: vtk: always assume combined output
The ASCII VTK output really shouldn't be used in the first place.
But it's usable as a quick way to look at data for small runs. It's
not scalable because it writes from a single proc, and ASCII is slow.

OTOH, if you don't have HDF5, this is what you get.

Both this VTK output as well as the HDF5 only are the only ones using
the globally assembled files (output_combine), and the HDF5 should also
go away except that it's currently used for the tests, so we keep it
around until someone gets to fixing the tests to do something better.