cbe: Officially builds on cell.
Drifting segfault Heisenbug when it actually runs.
Segfault is occuring during memcpy moving a new
active block into place.

My current suspicion is related to the spe tracking.
At the very least, the spe tracking/management needs
to be improved, so that's the next task. Heisenbug
should be easier to find when that's finished.