cbe: Skeleton Framework done.
Much of this will probably end up being modified
as the code progresses, but in the end the method of
handling blocks and the spu tracking will likely
stay the same.

At this point we've just got a ton of print statements.
If the message counts come back as less then they should,
it's likely that the ppe is running too fast and the buffers
aren't getting flushed before exit. I know, haha, the ppe is
running too fast. Deal with it.