cbe: Added a lot of backend for block decomposition and particle streaming.
It's unlikely this will actually work. It's a little bit difficult
to emulate the actual spu particle pushed with the way I have it
set up right now. In the future it would probably be a good idea
to create a nice emulation environment, but as the code becomes
more complicated this becomes somewhat difficult. I've already
been bitten once by using command which Kai did not fully
implement.

Expected another commit from fishercat soon with lots of typo and
bug fixes.