cbe: Added in some field interface.
Because the fields have to be copied into caches for offloading
to the spes anyway, it doesn't make much sense to have a seperate
type for global (on single proc) fields. Provided I eventually
modify the code to have a global precision flag, we'll just use
the C fields stuff and have the block decomposition stuff worked
into cbe_blocks.c.

Moving to fishercat for testing.