cbe: Added field caching and am dma-ing the fields into the spe.
Soooo, this seems to work well. But I think I broke the world.
I'm essentially bringing in 64KiB of data at the start of the
push kernel, and stalling the thread until it gets there. Yet,
contrary to what one would expect, this *decreases* the execution
time. This should not be.

Honestly, I think it's an indication that the task scheduling procedure
is in desperate need of improvement. I can't think of any other
reason why the addition of the fields would improve performance.