cbe: modified cbe particle type to match the new patch structure kai is developing.
At this point there are two issues, alignment and padding, which are preventing
me from discarding the cbe specific particles in favor of using the c base. The
doxygen file note at the end of psc_particles_cbe.h has a more in depth discussion
of these issues.

WARNING: compile not attempted, probably many post-merge issues still to be resolved.
Do not branch from here and expect a working code.