cbe: Major reworking to remove the block system in favor of the new patches.
Basically, the new sub-domain system Kai is setting up
removes the need for a seperate domain decomposition into
CBE blocks. Along the way, I made some improvements and cleaned up the code
a great deal.

A large amount of the ppu code for managing the queuing and whatnot
has been removed, and we're really stripped down to bare essentials now.
This raises some issues, notably that I'll have to make a bunch of changes
if we want the ppu to be doing anything while the spes run. I actually think
it may be possible, in the future, to have the ppu do the boundary exchanges on
completed blocks while the spes are finishing other domains. To implement this is,
however, a very conceptually complicated task. As such, the code is not
primed to accept this type of behaviour at the moment.

I assume there are some errors in the current code. Because I have broken
the emulation compatability, this needs to be moved to fishercat to test.