cbe: Split the particle pusher kernel. 2D is now xy, and yz has it's own file.
I don't really want to have a seperate kernel for each dimensionality. The big
problem is that if I want to integrate with the existing patch system and domain
decomposition then the SPU, as it's written, can **only** use the xy particle pusher.
Offloading onto the spu for the 2d pusher needs a flat plane of 32x32x1 data. To dma
it all in, that plane needs to be continuous in memory, which means xy.

I have also removed the old alf kernel which was still hanging around for some reason.

This commit doesn't have some of the backend work done, and hasn't been tested.