cbe: Performance tweak on the 2d kernel.
I was using static inline functions for some of the more commonly used
functions. This was a bad idea, as the functions were written to get pointers,
which meant (I think) that the values in registers were being written to the stack,
and then that address was being passed to the function.

Performance gain of ~20%.