bugfix/missing include: hacked a stdlib.h include into mrc_profile.h
There were a large number of compiler warnings popping up due to functions
needing stdlib.h. This file was being included from the without-papi profiling
header, but wasn't part of the papi profiling. To fix this properly we would
have to check that stdlib was being included by every file that needed it.
I have instead taken the quick fix route and added it to the papi section
in the profiling.