field_push + cbe: Hacked in a bit to make sure the subclass setup gets called.
Some modules (the cbe version in particular) need to have the subclass ops
'setup' and 'destroy' functions called, so we can just blindly forward it to
the bnd ops. At the moment, setup is forwarded to the ops before calling into
bnd, and destroy is after calling into bnd. This ordering may need to change.