psc/init: initialize particles on cell centers
It's not that we have to do it this way, but it avoids problems with
non-periodic b.c., and generally seems more consistent.

Unfortunately, however, the 2nd order pushers get the invariant index
wrong, and possibly the Fortran moment stuff, too, so we keep the
invariant coordinate as before for now.

It fixes a cuda test case, more by chance than anything.