libmrc/params: handle deprecated options
Initially, libmrc accepted options "as-is", ie., if the option is called
"xyz", it would look at --xyz on the command line. As multiple objects
may have the same command line options, this was then extended to also
handle the prefixed version, for example --my_object_name_xyz. When
printing --help, we only show the later version now, and we also
generally warn when using the non-prefixed version on the command line.