bugfix/api change: mrc_m1 and mrc_f3 _create overloads removed, mrc_f3_set_nr_comps removed.
The number of components for field data types cannot be set before the
object is created. For the two modified field types (m1 and f3) the overloaded
create function was allocating space for the '_comp_name' attribute, but because
nr_comp was not set space was only allocated for the default of 1 char pointer.
Later, when 'set_comp_name' is called, it attempts to free the memory at the pointers
in '_comp_name'. If attempting to set the name for a component greater than the default 1
then we've run off the end of the list '_comp_name' and are attempting to free the memory at some
random pointer.

This doesn't seem to cause a problem on most platforms, as that piece of memory we aren't supposed
to be touching seems to set to zero ( I think OSX, for example, periodically runs through and zeros
unalloced memory). On some systems, however, this results in a segfault and the bug becomes apparent.

I have standardized the mrc_m1 and mrc_f3 creation/setup process to that of the other datatypes.
The number of components should be set AFTER creation with set_param_int, but BEFORE setup. 'setup'
now handles allocating the '_comp_name' array.

NB: This change requires that 'setup' be called before 'set_comp_name'

FUNCTION REMOVED: mrc_f3_set_nr_comps
The change in setup/creation required removing mrc_f3_set_nr_comps. As written, f3_set_nr_comps
occured between setup and creation, but would again attempt to free random pieces of memory off
the end of a list. 'f3' was the only data type to have this function, so in the interests of making
the api consistent I have removed it rather in favor of set_param_int(..,"nr_comps",..) rather than
attempt to fix it.