psc/util: add Liang's python plotting utilities
These may be useful for others, e.g., the ability to calculate and
plot the flux function.

The most fundamental functions are

getfld: return the field data in 2D; the dimension can be specified like dims='xz' etc, (though default is xz). Specifying a small box to select data, as well as striding arrays, is allowed.
v2d: visualize the 2D data array
view2d: visualize the 2D data by specifying name of the field (basically getfld+v2d). This will be a most easy and fast command to use
v1d: visualize the sliced data of a 2D data array

Additionally, the function calcfld is useful for me to calculate quantities that are not provided in the PSC data directly, like jye(electron current in y), psi(magnetic flux, whose contours are magnetic field lines), uxe(electron flow velocity in x).

I personally use them in ipython or simply python. I also found it useful to add the directory where the script is located to the python path, so that I can import it wherever I am. It also requires h5py, matplotlib and numpy are properly installed in the system.