[mrc-v3] mrc_petsc_vec: add support for blocksizes and pre-existing petsc_vec
Right now this will only be needed in in mrc-v3, but that code is so
tightly wound around petsc that we need a way to implement creating mrc_fld,
and thus mrc_vec, with an existing petsc vector. Setting "petsc_vec" is currently
untested, and may not work.

Blocksize information may eventually move out to mrc_vec and be used in
libmrc scatter and gather implementations. (Big 'may')