[mrc-v3] libmrc: petsc integration: add methods to get naked petsc Vec
Not really happy with the implementation here, but we really need
to be able to get naked petsc vectors so they can be passed into
petsc functions.

get and put are strictly enforced within the bounds of setup-->destroy.

set is a bit tricky. To use the method, mrc_vec_set must be called with "petsc"
before the method was used. There needs to be a better solution to this.