[ggcm] mrc_io: fix mrc_fld I/O issues
We now again try to use write/read_f3/field methods if available and
no suitable write/read_m3 exists.

In the end, there should just be write_fld and nothing else, but now,
what used to be mrc_f3 looks like mrc_m3 with a single patch, in which
case we can still do I/O on it if the less generic f3 routine exists, so
we do that for compatibility / fix some tests.