[mrc-v3] Fixed [read,write]_fld so 1D flds write ghost points
This is a hacky way to have the crds io work, though it's still not
the proper way to write them (they get written as a bunch of highly
degenerate patches). Until, however, I figure out how to handle blocks
there's no point in writting a better function.

In the future it would be nice to add a parameter to the write_fld that
let's one choose how many ghost points to write.