[mrc-v3] Add coordinate and domain size setting to block_factories
This requires a major refactoring of how block_factories exist
in the domain. They are now member objects that persist for the domain's
lifetime, rather than ephemeral objects during setup. The take the dims setting
away from the domain, and take the crds_gen setting away from the crds.

gdims calls needed to be removed from the crds_gen because this isn't
defined for mb domains. Had to add a "dims" parameter to crds_gens to
account for this.

As a side-effect of all this the domain no longer has a trafo element.
That could change if it proves more convenient, but I like it this way
for now.