[ggcm] FIX: hack around hanging I/O server
This is not a proper fix, as I just skipped the
"wait_for_response", but it makes progress in
getting the buildbot tests to work again.

The issue is that there may be more than one mrc_io
talking to the same diag server and only one can tell
that to shut down and actually expect a response.