mrc_domain_multi: Force regeneration of the patch layout on read for parallel io
There are some fundamental issues with mrc_domain_multi "io". For
PSC checkpointing to work the patch numbers on each proc need to be
written and read, which (at this time) can only be done with a serial
io type where each proc writes its own bit. It is unreasonable, however,
to break all other multi domain io just to allow this.

This commit reads/writes the per proc patch info for serial io, and
regenerates the entire patch structure for parallel io. This is *not*
ideal, as it might silently break if someone tries to do psc checkpointing
with a parallel io type.

Reference results for tests have been updated to reflect that full patch
information is not being saved anymore.