io/fld : Don't register xdmf2_parallel io type, and test xdmf2 instead
xdmf2 was being used but not tested. xdmf2_parallel was being tested,
found eternally wanting, and never being used. I've moved the test so
it checks xdmf2 instead, and deregistered xdmf2_parallel. The code for
it hasn't been removed, but if nobody notices that it's gone we should
remove it all together.

Test inclusion references ticket:8