io : have xdmf_collective look at the correct dimension of slab constraints
When the m1 writing for xdmf_collective was written it seemed to assume
that a 1D field would only exist if there was only one spatial dimension
in the domain, so it always looked at the [0] component of slab_dims and
slab_offs for writing. That's not the case anymore, so I changed it to look
at the [dim] component. nr_ghost points is now being chosen properly,
so the ref results have to be updated to include the extra ghost cells.

Note that this is probably still being done incorrectly, see ticket:20