bits: make the KH checkpoint test a little more of a test
This still isn't a proper test, but it's closer. It writes a checkpoint
at 5 steps, runs to ten steps, then restarts using the checkpoint at 5 and
compared the data to the original run. Note there will be a bunch of
differences in h5dumps, notabaly UIDs and the fact that xdmf_serial reads/writes
the coordinate fields wrong.