[ggcm] fix: crd_nc did not have the right slab_dims
So this is all hacky anyway, but what happens is
that the node-centered coordinates get used as coordinates
in xdmf, and so they need to be written without ghost
points, and that's achieved by using the (ugly) slab_dims
feature. However, the limited (no-ghost) range was always
set on component 0, so that worked in the x direction, but
not y and z.

That's now fixed.