[ggcm] mrc_fld: don't write / read back _nd
This is rather ugly, because we can't control whether
a member object will be read back or not (it always will
be). But we're writing the mrc_fld data directly, so we
don't want to also write the underlying mrc_ndarray.

The workaround for now is to not list _nd as MRC_VAR_OBJ.

Also includes lots of reference result updates.