[ggcm] norm: add xxnorm, xxnorm0
xxnorm is the normalizing length scale (this length in m
will be 1 in code units). xxnorm0 is the I/O unit,
ie., the input/output is in term  of (xxnorm0 m).

In the typical OpenGGCM case, they are both equal to R_E,
which in fact means the conversion factor between input/output
in R_E and the internal length unit (also R_E) is 1.