fix psc_whistler.cxx to properly handle 3 species
Usually it's best not to specify a number of populations (last argument to
`setup_particles()`, so it'll automatically use the number of species
(a.k.a., particle kinds).

I didn't touch the `neutralizing_population`, but that may need to be fixed
as well. Generally speaking, it's good practice to use named constants
instead of hard coded numbers, e.g., replace "4" by "N_MY_KINDS", so it'll
automatically adjust if you change the number of particle kinds used.